rename predicates model_list => model and model => record
